requirements.txt
Flask==2.2.3
Flask-WTF==1.1.1
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
transformers==4.28.1
torch==2.0.1
spacy==3.5.2
nltk==3.8.1
python-dotenv==1.0.0
gunicorn==20.1.0
